+++ traework/main.go
@@ -273,7 +273,7 @@ type registrationCapability struct {
 }
 
 // version is injected at build time via -ldflags "-X main.version=...".
-var version = "0.1.19"
+var version = "0.1.20"
 
 func wbRegistration() registration {
 	return registration{